fix: move range value tests in right file
